fixed comment after review

+++ src/components/application_manager/include/application_manager/policies/policy_handler.h
@@ -317,7 +317,7 @@ class PolicyHandler :
    * @return if timeout was set then value in milliseconds greater zero
    * otherwise heart beat for specific application isn't set
    */
-  uint16_t HeartBeatTimeout(const std::string& app_id) const;
+  uint32_t HeartBeatTimeout(const std::string& app_id) const;
 
   /**
    * @brief Returns URL for querying list of remote apps
