+++ src/components/application_manager/src/policies/policy_handler.cc
@@ -1252,7 +1252,7 @@ bool PolicyHandler::CheckSystemAction(
   return false;
 }
 
-uint16_t PolicyHandler::HeartBeatTimeout(const std::string& app_id) const {
+uint32_t PolicyHandler::HeartBeatTimeout(const std::string& app_id) const {
   POLICY_LIB_CHECK(0);
   return policy_manager_->HeartBeatTimeout(app_id);
 }
