+++ src/components/policy/src/policy/include/policy/cache_manager.h
@@ -558,7 +558,7 @@ class CacheManager : public CacheManagerInterface {
    * @return if timeout was set then value in milliseconds greater zero
    * otherwise heart beat for specific application isn't set
    */
-  uint16_t HeartBeatTimeout(const std::string& app_id) const;
+  uint32_t HeartBeatTimeout(const std::string& app_id) const;
 
   /**
    * @brief Allows to generate hash from the specified string.
