+++ src/components/policy/src/policy/include/policy/cache_manager_interface.h
@@ -551,7 +551,7 @@ class CacheManagerInterface {
    * @return if timeout was set then value in milliseconds greater zero
    * otherwise heart beat for specific application isn't set
    */
-  virtual uint16_t HeartBeatTimeout(const std::string& app_id) const = 0;
+  virtual uint32_t HeartBeatTimeout(const std::string& app_id) const = 0;
 
   /**
    * @brief Resets all calculated permissions in cache
