+++ src/components/policy/src/policy/include/policy/policy_manager.h
@@ -395,7 +395,7 @@ class PolicyManager : public usage_statistics::StatisticsManager {
      * @return if timeout was set then value in milliseconds greater zero
      * otherwise heart beat for specific application isn't set
      */
-    virtual uint16_t HeartBeatTimeout(const std::string& app_id) const = 0;
+    virtual uint32_t HeartBeatTimeout(const std::string& app_id) const = 0;
 
     /**
      * @brief SaveUpdateStatusRequired alows to save update status.
