+++ src/components/policy/src/policy/policy_table/table_struct/types.h
@@ -93,7 +93,7 @@ struct ApplicationParams : PolicyBase {
     Optional< AppHMITypes > AppHMIType;
     Optional< RequestTypes > RequestType;
     Optional< Integer<uint16_t, 0, 65225> > memory_kb;
-    Optional< Integer<uint16_t, 0, 65225> > heart_beat_timeout_ms;
+    Optional< Integer<uint32_t, 0, 65225> > heart_beat_timeout_ms;
     Optional< String<0, 255> > certificate;
   public:
     ApplicationParams();
